Revert^2 "Audio CAP: Address ANAPIC comments, Part 3."
See ag/30063274.

Combine AudioPolicyForcedConfig with AudioPolicyForceUse and
provide a more structured way for setting values. Thus, instead
of a loosely coupled pair of values from the enum
'AudioPolicyForceUse' and the associated enum
'AudioPolicyForcedConfig', there is now a single union
'AudioPolicyForceUse'. The tag of this union replaces the enum
named the same, while the value of the union replaces
'AudioPolicyForcedConfig'. Possible values for uses got limited
to the pairs that actually make sense.

For consistency, `forceUse` field was removed from
'AudioHalCapCriterionV2::ForceConfigForUse'. Instead, the tag
from the first element of 'values' is used for specifying
the use case.

Bug: 364310317
Test: m
Change-Id: Iaf21224136837cc524586772dde5b72152e2bcc9